Refactor distinct values metrics into separate files
Split new metrics from column_distinct_values.py into:
- column_distinct_values_not_in_set.py: ColumnDistinctValuesNotInSetCount, ColumnDistinctValuesNotInSet
- column_distinct_values_missing_from_set.py: ColumnDistinctValuesMissingFromSetCount, ColumnDistinctValuesMissingFromSet

Also replaced hasattr checks with version comparison (_SQLALCHEMY_1_4_OR_GREATER)
for clearer SQLAlchemy compatibility handling.
