+++ great_expectations/expectations/metrics/column_aggregate_metrics/column_distinct_values.py
@@ -5,8 +5,12 @@ from typing import TYPE_CHECKING, Any, Dict, List, Optional, Set
 
 from dateutil.parser import parse
 
+from great_expectations.compatibility.not_imported import is_version_greater_or_equal
 from great_expectations.compatibility.pyspark import (
     functions as F,
+)
+from great_expectations.compatibility.sqlalchemy import (
+    __version__ as SQLALCHEMY_VERSION,
 )
 from great_expectations.compatibility.sqlalchemy import sqlalchemy as sa
 from great_expectations.compatibility.typing_extensions import override
@@ -32,6 +36,11 @@ if TYPE_CHECKING:
     from great_expectations.expectations.expectation_configuration import (
         ExpectationConfiguration,
     )
+
+# SQLAlchemy 1.4+ uses is_not() instead of isnot()
+_SQLALCHEMY_1_4_OR_GREATER = SQLALCHEMY_VERSION is not None and is_version_greater_or_equal(
+    SQLALCHEMY_VERSION, "1.4.0"
+)
 
 
 def _coerce_value_set_to_column_type(column_set: Set[Any], value_set: List[Any]) -> Set[Any]:
@@ -119,7 +128,7 @@ class ColumnDistinctValues(ColumnAggregateMetricProvider):
         column: sqlalchemy.ColumnClause = sa.column(column_name)
 
         distinct_values: List[sqlalchemy.Row]
-        if hasattr(column, "is_not"):
+        if _SQLALCHEMY_1_4_OR_GREATER:
             distinct_values = execution_engine.execute_query(  # type: ignore[assignment] # FIXME CoP
                 sa.select(column).where(column.is_not(None)).distinct().select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
             ).fetchall()
@@ -244,415 +253,3 @@ class ColumnDistinctValuesCountUnderThreshold(ColumnAggregateMetricProvider):
                 metric_value_kwargs=None,
             )
         return dependencies
-
-
-class ColumnDistinctValuesNotInSetCount(ColumnAggregateMetricProvider):
-    """Metric that returns count of column values NOT in the expected set.
-
-    Used for expect_column_distinct_values_to_be_in_set to determine pass/fail
-    without fetching all distinct values.
-    """
-
-    metric_name = "column.distinct_values.not_in_set.count"
-    value_keys = ("value_set",)
-
-    @column_aggregate_value(engine=PandasExecutionEngine)
-    def _pandas(cls, column: pd.Series, value_set: List[Any], **kwargs) -> int:
-        column_set = set(column.dropna().unique())
-        expected_set = _coerce_value_set_to_column_type(column_set, value_set)
-        return len(column_set - expected_set)
-
-    @metric_value(engine=SqlAlchemyExecutionEngine)
-    def _sqlalchemy(
-        cls,
-        execution_engine: SqlAlchemyExecutionEngine,
-        metric_domain_kwargs: Dict[str, str],
-        metric_value_kwargs: Dict[str, Any],
-        **kwargs,
-    ) -> int:
-        """Count distinct values in column that are NOT in the expected set."""
-        value_set = _coerce_value_set_for_sql(metric_value_kwargs.get("value_set", []))
-
-        selectable: sqlalchemy.Selectable
-        accessor_domain_kwargs: Dict[str, str]
-        (
-            selectable,
-            _,
-            accessor_domain_kwargs,
-        ) = execution_engine.get_compute_domain(metric_domain_kwargs, MetricDomainTypes.COLUMN)
-        column_name: str = accessor_domain_kwargs["column"]
-        column: sqlalchemy.ColumnClause = sa.column(column_name)
-
-        # Count distinct values NOT in the provided set
-        if value_set:
-            if hasattr(column, "is_not"):
-                query = (
-                    sa.select(sa.func.count(sa.distinct(column)))
-                    .where(column.is_not(None))
-                    .where(column.notin_(value_set))
-                    .select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-                )
-            else:
-                query = (
-                    sa.select(sa.func.count(sa.distinct(column)))
-                    .where(column.isnot(None))
-                    .where(column.notin_(value_set))
-                    .select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-                )
-        # Empty value_set means all non-null values are violations
-        elif hasattr(column, "is_not"):
-            query = (
-                sa.select(sa.func.count(sa.distinct(column)))
-                .where(column.is_not(None))
-                .select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-            )
-        else:
-            query = (
-                sa.select(sa.func.count(sa.distinct(column)))
-                .where(column.isnot(None))
-                .select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-            )
-
-        result = execution_engine.execute_query(query).scalar()
-        return result or 0
-
-    @metric_value(engine=SparkDFExecutionEngine)
-    def _spark(
-        cls,
-        execution_engine: SparkDFExecutionEngine,
-        metric_domain_kwargs: Dict[str, str],
-        metric_value_kwargs: Dict[str, Any],
-        **kwargs,
-    ) -> int:
-        """Count distinct values in column that are NOT in the expected set."""
-        value_set = metric_value_kwargs.get("value_set", [])
-
-        df: pyspark.DataFrame
-        accessor_domain_kwargs: Dict[str, str]
-        (
-            df,
-            _,
-            accessor_domain_kwargs,
-        ) = execution_engine.get_compute_domain(metric_domain_kwargs, MetricDomainTypes.COLUMN)
-        column_name: str = accessor_domain_kwargs["column"]
-
-        # Filter to values NOT in the set and count distinct
-        filtered_df = df.where(F.col(column_name).isNotNull())
-        if value_set:
-            filtered_df = filtered_df.where(~F.col(column_name).isin(value_set))
-
-        result = filtered_df.select(F.countDistinct(F.col(column_name))).collect()[0][0]
-        return result or 0
-
-
-class ColumnDistinctValuesNotInSet(ColumnAggregateMetricProvider):
-    """Metric that returns sample of column values NOT in the expected set.
-
-    Used for expect_column_distinct_values_to_be_in_set to report violations
-    without fetching all distinct values.
-    """
-
-    metric_name = "column.distinct_values.not_in_set"
-    value_keys = ("value_set", "limit")
-
-    @column_aggregate_value(engine=PandasExecutionEngine)
-    def _pandas(
-        cls, column: pd.Series, value_set: List[Any], limit: int = 20, **kwargs
-    ) -> List[Any]:
-        column_set = set(column.dropna().unique())
-        expected_set = _coerce_value_set_to_column_type(column_set, value_set)
-        not_in_set = list(column_set - expected_set)
-        # Sort for deterministic results, handling mixed types gracefully
-        try:
-            not_in_set = sorted(not_in_set)
-        except TypeError:
-            pass  # Mixed types can't be sorted, return unsorted
-        return not_in_set[:limit]
-
-    @metric_value(engine=SqlAlchemyExecutionEngine)
-    def _sqlalchemy(
-        cls,
-        execution_engine: SqlAlchemyExecutionEngine,
-        metric_domain_kwargs: Dict[str, str],
-        metric_value_kwargs: Dict[str, Any],
-        **kwargs,
-    ) -> List[Any]:
-        """Return sample of distinct values in column that are NOT in the expected set."""
-        value_set = _coerce_value_set_for_sql(metric_value_kwargs.get("value_set", []))
-        limit = metric_value_kwargs.get("limit", 20)
-
-        selectable: sqlalchemy.Selectable
-        accessor_domain_kwargs: Dict[str, str]
-        (
-            selectable,
-            _,
-            accessor_domain_kwargs,
-        ) = execution_engine.get_compute_domain(metric_domain_kwargs, MetricDomainTypes.COLUMN)
-        column_name: str = accessor_domain_kwargs["column"]
-        column: sqlalchemy.ColumnClause = sa.column(column_name)
-
-        # Get distinct values NOT in the provided set (with limit)
-        if value_set:
-            if hasattr(column, "is_not"):
-                query = (
-                    sa.select(column)
-                    .where(column.is_not(None))
-                    .where(column.notin_(value_set))
-                    .distinct()
-                    .limit(limit)
-                    .select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-                )
-            else:
-                query = (
-                    sa.select(column)
-                    .where(column.isnot(None))
-                    .where(column.notin_(value_set))
-                    .distinct()
-                    .limit(limit)
-                    .select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-                )
-        # Empty value_set means all non-null values are violations
-        elif hasattr(column, "is_not"):
-            query = (
-                sa.select(column)
-                .where(column.is_not(None))
-                .distinct()
-                .limit(limit)
-                .select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-            )
-        else:
-            query = (
-                sa.select(column)
-                .where(column.isnot(None))
-                .distinct()
-                .limit(limit)
-                .select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-            )
-
-        results = execution_engine.execute_query(query).fetchall()
-        return [row[0] for row in results]
-
-    @metric_value(engine=SparkDFExecutionEngine)
-    def _spark(
-        cls,
-        execution_engine: SparkDFExecutionEngine,
-        metric_domain_kwargs: Dict[str, str],
-        metric_value_kwargs: Dict[str, Any],
-        **kwargs,
-    ) -> List[Any]:
-        """Return sample of distinct values in column that are NOT in the expected set."""
-        value_set = metric_value_kwargs.get("value_set", [])
-        limit = metric_value_kwargs.get("limit", 20)
-
-        df: pyspark.DataFrame
-        accessor_domain_kwargs: Dict[str, str]
-        (
-            df,
-            _,
-            accessor_domain_kwargs,
-        ) = execution_engine.get_compute_domain(metric_domain_kwargs, MetricDomainTypes.COLUMN)
-        column_name: str = accessor_domain_kwargs["column"]
-
-        # Filter to values NOT in the set
-        filtered_df = df.where(F.col(column_name).isNotNull())
-        if value_set:
-            filtered_df = filtered_df.where(~F.col(column_name).isin(value_set))
-
-        # Get distinct values with limit
-        results = (
-            filtered_df.select(F.col(column_name))
-            .distinct()
-            .limit(limit)
-            .rdd.flatMap(lambda x: x)
-            .collect()
-        )
-        return list(results)
-
-
-class ColumnDistinctValuesMissingFromSetCount(ColumnAggregateMetricProvider):
-    """Metric that returns count of expected values missing from the column.
-
-    Used for expect_column_distinct_values_to_contain_set to determine pass/fail
-    without fetching all distinct values.
-    """
-
-    metric_name = "column.distinct_values.missing_from_set.count"
-    value_keys = ("value_set",)
-
-    @column_aggregate_value(engine=PandasExecutionEngine)
-    def _pandas(cls, column: pd.Series, value_set: List[Any], **kwargs) -> int:
-        column_set = set(column.dropna().unique())
-        expected_set = _coerce_value_set_to_column_type(column_set, value_set)
-        missing = expected_set - column_set
-        return len(missing)
-
-    @metric_value(engine=SqlAlchemyExecutionEngine)
-    def _sqlalchemy(
-        cls,
-        execution_engine: SqlAlchemyExecutionEngine,
-        metric_domain_kwargs: Dict[str, str],
-        metric_value_kwargs: Dict[str, Any],
-        **kwargs,
-    ) -> int:
-        """Count values in the expected set that are missing from the column."""
-        value_set = _coerce_value_set_for_sql(metric_value_kwargs.get("value_set", []))
-        if not value_set:
-            return 0
-
-        selectable: sqlalchemy.Selectable
-        accessor_domain_kwargs: Dict[str, str]
-        (
-            selectable,
-            _,
-            accessor_domain_kwargs,
-        ) = execution_engine.get_compute_domain(metric_domain_kwargs, MetricDomainTypes.COLUMN)
-        column_name: str = accessor_domain_kwargs["column"]
-        column: sqlalchemy.ColumnClause = sa.column(column_name)
-
-        # Count how many expected values exist in the column
-        if hasattr(column, "is_not"):
-            query = (
-                sa.select(sa.func.count(sa.distinct(column)))
-                .where(column.is_not(None))
-                .where(column.in_(value_set))
-                .select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-            )
-        else:
-            query = (
-                sa.select(sa.func.count(sa.distinct(column)))
-                .where(column.isnot(None))
-                .where(column.in_(value_set))
-                .select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-            )
-
-        found_count = execution_engine.execute_query(query).scalar() or 0
-        return len(value_set) - found_count
-
-    @metric_value(engine=SparkDFExecutionEngine)
-    def _spark(
-        cls,
-        execution_engine: SparkDFExecutionEngine,
-        metric_domain_kwargs: Dict[str, str],
-        metric_value_kwargs: Dict[str, Any],
-        **kwargs,
-    ) -> int:
-        """Count values in the expected set that are missing from the column."""
-        value_set = metric_value_kwargs.get("value_set", [])
-        if not value_set:
-            return 0
-
-        df: pyspark.DataFrame
-        accessor_domain_kwargs: Dict[str, str]
-        (
-            df,
-            _,
-            accessor_domain_kwargs,
-        ) = execution_engine.get_compute_domain(metric_domain_kwargs, MetricDomainTypes.COLUMN)
-        column_name: str = accessor_domain_kwargs["column"]
-
-        # Count how many expected values exist in the column
-        found_count = (
-            (
-                df.where(F.col(column_name).isNotNull())
-                .where(F.col(column_name).isin(value_set))
-                .select(F.countDistinct(F.col(column_name)))
-                .collect()[0][0]
-            )
-            or 0
-        )
-
-        return len(value_set) - found_count
-
-
-class ColumnDistinctValuesMissingFromSet(ColumnAggregateMetricProvider):
-    """Metric that returns values in the expected set that are missing from the column.
-
-    Used for expect_column_distinct_values_to_contain_set to check which
-    required values are not present in the column.
-    """
-
-    metric_name = "column.distinct_values.missing_from_set"
-    value_keys = ("value_set", "limit")
-
-    @column_aggregate_value(engine=PandasExecutionEngine)
-    def _pandas(
-        cls, column: pd.Series, value_set: List[Any], limit: int = 20, **kwargs
-    ) -> List[Any]:
-        column_set = set(column.dropna().unique())
-        expected_set = _coerce_value_set_to_column_type(column_set, value_set)
-        missing = list(expected_set - column_set)
-        return missing[:limit]
-
-    @metric_value(engine=SqlAlchemyExecutionEngine)
-    def _sqlalchemy(
-        cls,
-        execution_engine: SqlAlchemyExecutionEngine,
-        metric_domain_kwargs: Dict[str, str],
-        metric_value_kwargs: Dict[str, Any],
-        **kwargs,
-    ) -> List[Any]:
-        """Return values in the expected set that are missing from the column."""
-        value_set = _coerce_value_set_for_sql(metric_value_kwargs.get("value_set", []))
-        limit = metric_value_kwargs.get("limit", 20)
-
-        selectable: sqlalchemy.Selectable
-        accessor_domain_kwargs: Dict[str, str]
-        (
-            selectable,
-            _,
-            accessor_domain_kwargs,
-        ) = execution_engine.get_compute_domain(metric_domain_kwargs, MetricDomainTypes.COLUMN)
-        column_name: str = accessor_domain_kwargs["column"]
-        column: sqlalchemy.ColumnClause = sa.column(column_name)
-
-        # Get distinct values in the column
-        if hasattr(column, "is_not"):
-            column_values_query = (
-                sa.select(column).where(column.is_not(None)).distinct().select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-            )
-        else:
-            column_values_query = (
-                sa.select(column).where(column.isnot(None)).distinct().select_from(selectable)  # type: ignore[arg-type] # FIXME CoP
-            )
-
-        column_values_result = execution_engine.execute_query(column_values_query).fetchall()
-        column_values_set = {row[0] for row in column_values_result}
-
-        # Find missing values
-        missing = [v for v in value_set if v not in column_values_set]
-        return missing[:limit]
-
-    @metric_value(engine=SparkDFExecutionEngine)
-    def _spark(
-        cls,
-        execution_engine: SparkDFExecutionEngine,
-        metric_domain_kwargs: Dict[str, str],
-        metric_value_kwargs: Dict[str, Any],
-        **kwargs,
-    ) -> List[Any]:
-        """Return values in the expected set that are missing from the column."""
-        value_set = metric_value_kwargs.get("value_set", [])
-        limit = metric_value_kwargs.get("limit", 20)
-
-        df: pyspark.DataFrame
-        accessor_domain_kwargs: Dict[str, str]
-        (
-            df,
-            _,
-            accessor_domain_kwargs,
-        ) = execution_engine.get_compute_domain(metric_domain_kwargs, MetricDomainTypes.COLUMN)
-        column_name: str = accessor_domain_kwargs["column"]
-
-        # Get distinct values in the column
-        column_values = (
-            df.select(F.col(column_name))
-            .where(F.col(column_name).isNotNull())
-            .distinct()
-            .rdd.flatMap(lambda x: x)
-            .collect()
-        )
-        column_values_set = set(column_values)
-
-        # Find missing values
-        missing = [v for v in value_set if v not in column_values_set]
-        return missing[:limit]
